refactor: protected api routes

+++ src/app/api/jobs/route.ts
@@ -1,10 +1,20 @@
 import { NextResponse } from "next/server";
 import { register } from "@/instrumentation";
 import { Job } from "../../../../db/schema";
+import { verifyFirebaseToken } from "@/lib/verifyFirebaseToken";
 
-export async function GET() {
+export async function GET(req: Request) {
   try {
     await register();
+    const authHeader = req.headers.get("Authorization");
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return NextResponse.json({ error: "Missing token" }, { status: 401 });
+    }
+    const token = authHeader.split(" ")[1];
+    const decodedToken = await verifyFirebaseToken(token);
+    if (!decodedToken) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
     const jobs = await Job.find().sort({ createdAt: -1 });
     return NextResponse.json({ jobs });
   } catch (error) {
@@ -19,6 +29,15 @@ export async function POST(req: Request) {
     const { newJob } = body;
 
     await register();
+    const authHeader = req.headers.get("Authorization");
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return NextResponse.json({ error: "Missing token" }, { status: 401 });
+    }
+    const token = authHeader.split(" ")[1];
+    const decodedToken = await verifyFirebaseToken(token);
+    if (!decodedToken) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
     const job = new Job(newJob);
     await job.save();
 
@@ -35,6 +54,15 @@ export async function PATCH(req: Request) {
     const { jobId, updates } = body;
 
     await register();
+    const authHeader = req.headers.get("Authorization");
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return NextResponse.json({ error: "Missing token" }, { status: 401 });
+    }
+    const token = authHeader.split(" ")[1];
+    const decodedToken = await verifyFirebaseToken(token);
+    if (!decodedToken) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
     const updatedJob = await Job.findByIdAndUpdate(jobId, updates, { new: true });
 
     if (!updatedJob) {
@@ -54,6 +82,15 @@ export async function DELETE(req: Request) {
     const { jobId } = body;
 
     await register();
+    const authHeader = req.headers.get("Authorization");
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return NextResponse.json({ error: "Missing token" }, { status: 401 });
+    }
+    const token = authHeader.split(" ")[1];
+    const decodedToken = await verifyFirebaseToken(token);
+    if (!decodedToken) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
     const deletedJob = await Job.findByIdAndDelete(jobId);
 
     if (!deletedJob) {
